# contracts/product-api/v2/openapi.json
{
  "components": {
    "schemas": {
      "BacktestCancellationResponse": {
        "additionalProperties": false,
        "properties": {
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "state",
          "revision"
        ],
        "title": "BacktestCancellationResponse",
        "type": "object"
      },
...
      "BacktestRunCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "dataset_binding_fingerprint": {
            "title": "Dataset Binding Fingerprint",
            "type": "string"
          },
          "execution_profile": {
            "$ref": "#/components/schemas/BacktestProfileReferenceDto"
          },
          "initial_account": {
            "$ref": "#/components/schemas/BacktestInitialAccountDto"
          },
          "market_product_configuration_fingerprint": {
            "title": "Market Product Configuration Fingerprint",
            "type": "string"
          },
          "portfolio_profile": {
            "$ref": "#/components/schemas/BacktestProfileReferenceDto"
          },
          "risk_profile": {
            "$ref": "#/components/schemas/BacktestProfileReferenceDto"
          },
          "runtime_options": {
            "$ref": "#/components/schemas/BacktestRuntimeOptionsDto",
            "default": {
              "ordered_fact_policy": "ORDERED_FACTS_V1"
            }
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "strategy_fingerprint": {
            "title": "Strategy Fingerprint",
            "type": "string"
          }
        },
        "required": [
          "strategy_fingerprint",
          "dataset_binding_fingerprint",
          "market_product_configuration_fingerprint",
          "portfolio_profile",
          "risk_profile",
          "execution_profile",
          "initial_account"
        ],
        "title": "BacktestRunCreateRequest",
        "type": "object"
      },
      "BacktestRunCreateResponse": {
        "additionalProperties": false,
        "properties": {
          "backtest_run_id": {
            "title": "Backtest Run Id",
            "type": "string"
          },
          "disposition": {
            "title": "Disposition",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "backtest_run_id",
          "state",
          "disposition"
        ],
        "title": "BacktestRunCreateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductErrorEnvelopeDto": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ProductErrorDto"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "error"
        ],
        "title": "ProductErrorEnvelopeDto",
        "type": "object"
      },
...
      "ResearchRunErrorEnvelopeDto": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ResearchRunErrorDto"
          }
        },
        "required": [
          "error"
        ],
        "title": "ResearchRunErrorEnvelopeDto",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OnlyAlpha Product API",
    "version": "2"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/backtest/runs": {
      "post": {
        "operationId": "create_run_api_v2_backtest_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Run",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/v2/backtest/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v2_backtest_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Run",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/v2/backtest/runs/{run_id}/cancellation": {
      "post": {
        "operationId": "cancel_run_api_v2_backtest_runs__run_id__cancellation_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestCancellationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "backtest"
        ]
      }
    },
...
    "/api/v2/research/runs": {
      "get": {
        "operationId": "list_runs_api_v2_research_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunPageDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Runs",
        "tags": [
          "research-runs"
        ]
      },
      "post": {
        "operationId": "submit_run_api_v2_research_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitResearchRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResearchRunResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Idempotency-Key": {
                "description": "Exact admitted Product Command identity",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Run",
        "tags": [
          "research-runs"
        ]
      }
    },
    "/api/v2/research/runs/commands": {
      "post": {
        "operationId": "submit_research_run_command_v2",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitResearchRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResearchRunResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Idempotency-Key": {
                "description": "Exact admitted Product Command identity",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Run Command",
        "tags": [
          "research-runs"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [],
          "owning_authority_references": [
            {
              "locator_kind": "UUID4",
              "reference_kind": "RESEARCH_RUN",
              "reference_schema_version": 1,
              "response_field": "run.run_id"
            }
          ],
          "product_command_id_transport": {
            "in": "header",
            "name": "Idempotency-Key"
          },
          "recovery_class": "IDEMPOTENT_COMMAND",
          "requires_product_command_id": true,
          "schema_version": 1,
          "tool_class": "RESEARCH_RUN_SUBMIT"
        }
      }
    },
    "/api/v2/research/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v2_research_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string",
              "x-onlyalpha-reference-kind": "RESEARCH_RUN",
              "x-onlyalpha-reference-locator-kind": "UUID4",
              "x-onlyalpha-reference-schema-version": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Run",
        "tags": [
          "research-runs"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "run_id"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "UUID4",
              "reference_kind": "RESEARCH_RUN",
              "reference_schema_version": 1,
              "response_field": "run_id"
            }
          ],
          "recovery_class": "MUTABLE_OBSERVATION_QUERY",
          "requires_product_command_id": false,
          "schema_version": 1,
          "tool_class": "RESEARCH_RUN_QUERY"
        }
      }
    },
    "/api/v2/research/runs/{run_id}/cancellation": {
      "post": {
        "operationId": "cancel_run_api_v2_research_runs__run_id__cancellation_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "research-runs"
        ]
      }
    },
...
  }
}